requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.7
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
